refactor: ProjectList 스타일 파일 이동
